Big refactor - part 1
