+++ DataLayer/DataLayer.csproj
@@ -17,7 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\CommonCache\CommonCache.csproj" />
-    <ProjectReference Include="..\DataAuthorize\DataAuthorize.csproj" />
+    <ProjectReference Include="..\DataKeyParts\DataKeyParts.csproj" />
     <ProjectReference Include="..\PermissionParts\PermissionParts.csproj" />
   </ItemGroup>
