+++ DataLayer/EfCode/Configurations/ConfigExtensions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using DataAuthorize;
+using DataKeyParts;
 using DataLayer.ExtraAuthClasses;
 using DataLayer.MultiTenantClasses;
 using Microsoft.EntityFrameworkCore;
