+++ DataLayer/MultiTenantClasses/ShopStock.cs
@@ -2,7 +2,7 @@
 // Licensed under MIT license. See License.txt in the project root for license information.
 
 using System.ComponentModel.DataAnnotations.Schema;
-using DataAuthorize;
+using DataKeyParts;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace DataLayer.MultiTenantClasses
