+++ PermissionAccessControl2.sln
@@ -25,6 +25,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataKeyParts", "DataKeyParts\DataKeyParts.csproj", "{B76F79A2-A02E-4879-B2D4-6DC20483250F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -63,6 +65,10 @@ Global
 		{F2A3A002-D1A3-4BC4-AA6B-B11EFEEE9F87}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F2A3A002-D1A3-4BC4-AA6B-B11EFEEE9F87}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F2A3A002-D1A3-4BC4-AA6B-B11EFEEE9F87}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B76F79A2-A02E-4879-B2D4-6DC20483250F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B76F79A2-A02E-4879-B2D4-6DC20483250F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B76F79A2-A02E-4879-B2D4-6DC20483250F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B76F79A2-A02E-4879-B2D4-6DC20483250F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
