+++ PermissionAccessControl2/PermissionAccessControl2.csproj
@@ -22,6 +22,7 @@
   <ItemGroup>
     <ProjectReference Include="..\CommonCache\CommonCache.csproj" />
     <ProjectReference Include="..\DataAuthorize\DataAuthorize.csproj" />
+    <ProjectReference Include="..\DataKeyParts\DataKeyParts.csproj" />
     <ProjectReference Include="..\DataLayer\DataLayer.csproj" />
     <ProjectReference Include="..\FeatureAuthorize\FeatureAuthorize.csproj" />
     <ProjectReference Include="..\PermissionParts\PermissionParts.csproj" />
