+++ PermissionAccessControl2/Startup.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using CommonCache;
-using DataAuthorize;
+using DataKeyParts;
 using DataLayer.EfCode;
 using FeatureAuthorize;
 using FeatureAuthorize.PolicyCode;
@@ -19,6 +19,7 @@ using PermissionAccessControl2.Data;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using ServiceLayer.AppStart;
+using ServiceLayer.AuthoriseSetup;
 using ServiceLayer.CodeCalledInStartup;
 using ServiceLayer.UserServices;
 using Swashbuckle.AspNetCore.Swagger;
