+++ PermissionAccessControl2/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model DataAuthorize.DemoSetupOptions
+﻿@model DataKeyParts.DemoSetupOptions
 @{
     ViewData["Title"] = "Home Page";
 }
@@ -22,9 +22,9 @@
 <div>
     <h5>The setup from the appsetting.json DemoSetup section is:</h5>
     <ul>
-        <li><strong>@nameof(DataAuthorize.DemoSetupOptions.DatabaseSetup)</strong>: @Model.DatabaseSetup</li>
-        <li><strong>@nameof(DataAuthorize.DemoSetupOptions.CreateAndSeed)</strong>: @Model.CreateAndSeed</li>
-        <li><strong>@nameof(DataAuthorize.DemoSetupOptions.AuthCookieVersion)</strong>: @Model.AuthCookieVersion</li>
+        <li><strong>@nameof(DataKeyParts.DemoSetupOptions.DatabaseSetup)</strong>: @Model.DatabaseSetup</li>
+        <li><strong>@nameof(DataKeyParts.DemoSetupOptions.CreateAndSeed)</strong>: @Model.CreateAndSeed</li>
+        <li><strong>@nameof(DataKeyParts.DemoSetupOptions.AuthCookieVersion)</strong>: @Model.AuthCookieVersion</li>
     </ul>
 </div>
