+++ ServiceLayer/AuthoriseSetup/AuthCookieValidateImpersonation.cs
@@ -6,64 +6,45 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using CommonCache;
 using DataAuthorize;
+using DataKeyParts;
 using DataLayer.EfCode;
 using FeatureAuthorize;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.DataProtection;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using ServiceLayer.CodeCalledInStartup;
 using ServiceLayer.UserImpersonation.Concrete.Internal;
 
-namespace ServiceLayer.AuthCookieVersions
+namespace ServiceLayer.AuthoriseSetup
 {
     /// <summary>
     /// This version provides:
     /// - Adds Permissions to the user's claims.
     /// - Adds DataKey to the user's claims
-    /// - AND allows for user impersonation
-    /// - AND the user's claims are updated if there is a change in the roles/datakey information
+    /// - AND user impersonation
     /// </summary>
-    public class AuthCookieValidateEverything : IAuthCookieValidate
+    public class AuthCookieValidateImpersonation : IAuthCookieValidate
     {
-        /// <summary>
-        /// This will set up the user's feature permissions if either of the following states are found
-        /// - The current claims doesn't have the PackedPermissionClaimType. This happens when someone logs in.
-        /// - If the LastPermissionsUpdatedClaimType is missing (null) or is a lower number that is stored in the TimeStore cache.
-        /// It will also add a HierarchicalKeyClaimName claim with the user's data key if not present.
-        /// </summary>
-        /// <param name="context"></param>
-        /// <returns></returns>
         public async Task ValidateAsync(CookieValidatePrincipalContext context)
         {
-            var extraContext = context.HttpContext.RequestServices.GetRequiredService<ExtraAuthorizeDbContext>();
-            var protectionProvider = context.HttpContext.RequestServices.GetService<IDataProtectionProvider>();
-            var authChanges = new AuthChanges();
-
             var originalClaims = context.Principal.Claims.ToList();
+            var protectionProvider = context.HttpContext.RequestServices.GetService<IDataProtectionProvider>();
             var impHandler = new ImpersonationHandler(context.HttpContext, protectionProvider, originalClaims);
             
             var newClaims = new List<Claim>();
             if (originalClaims.All(x => x.Type != PermissionConstants.PackedPermissionClaimType) ||
-                impHandler.ImpersonationChange ||
-                authChanges.IsOutOfDateOrMissing(AuthChangesConsts.FeatureCacheKey, 
-                    originalClaims.SingleOrDefault(x => x.Type == PermissionConstants.LastPermissionsUpdatedClaimType)?.Value,
-                    extraContext))
+                impHandler.ImpersonationChange)
             {
+                //There is no PackedPermissionClaimType or there was a change in the impersonation state
+                
+                var extraContext = context.HttpContext.RequestServices.GetRequiredService<ExtraAuthorizeDbContext>();
                 var rtoPCalcer = new CalcAllowedPermissions(extraContext);
                 var dataKeyCalc = new CalcDataKey(extraContext);
 
-                //Handle the feature permissions
-                var permissionUserId = impHandler.GetUserIdForWorkingOutPermissions();
-                newClaims.AddRange(await BuildFeatureClaimsAsync(permissionUserId, rtoPCalcer));
-
-                //Handle the DataKey
-                var datakeyUserId = impHandler.GetUserIdForWorkingDataKey();
-                newClaims.AddRange(BuildDataClaims(datakeyUserId, dataKeyCalc));
-
-                //Something has changed so we replace the current ClaimsPrincipal with a new one
+                var userId = impHandler.GetUserIdForWorkingDataKey();
+                newClaims.AddRange(await BuildFeatureClaimsAsync(userId, rtoPCalcer));
+                newClaims.AddRange(BuildDataClaims(userId, dataKeyCalc));
 
                 newClaims.AddRange(RemoveUpdatedClaimsFromOriginalClaims(originalClaims, newClaims)); //Copy over unchanged claims
                 impHandler.AddOrRemoveImpersonationClaim(newClaims);
