+++ ServiceLayer/AuthoriseSetup/AuthCookieValidatePermissionsOnly.cs
@@ -5,21 +5,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using DataAuthorize;
 using DataLayer.EfCode;
 using FeatureAuthorize;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.DependencyInjection;
 using ServiceLayer.CodeCalledInStartup;
 
-namespace ServiceLayer.AuthCookieVersions
+namespace ServiceLayer.AuthoriseSetup
 {
     /// <summary>
-    /// This version provides:
+    /// This is the very basic version which only provides:
     /// - Adds Permissions to the user's claims.
-    /// - Adds DataKey to the user's claims
+    /// - NO DataKey
     /// </summary>
-    public class AuthCookieValidatePermissionsDataKey : IAuthCookieValidate
+    public class AuthCookieValidatePermissionsOnly : IAuthCookieValidate
     {
         public async Task ValidateAsync(CookieValidatePrincipalContext context)
         {
@@ -27,19 +26,14 @@ namespace ServiceLayer.AuthCookieVersions
                 return;
 
             //No permissions in the claims, so we need to add it. This is only happen once after the user has logged in
-            var extraContext = context.HttpContext.RequestServices.GetRequiredService<ExtraAuthorizeDbContext>();
-            var rtoPCalcer = new CalcAllowedPermissions(extraContext);
-            var dataKeyCalc = new CalcDataKey(extraContext);
+            var dbContext = context.HttpContext.RequestServices.GetRequiredService<ExtraAuthorizeDbContext>();
+            var rtoPCalcer = new CalcAllowedPermissions(dbContext);
 
             var claims = new List<Claim>();
             claims.AddRange(context.Principal.Claims); //Copy over existing claims
-            var userId = context.Principal.Claims.GetUserIdFromClaims();
             //Now calculate the Permissions Claim value and add it
             claims.Add(new Claim(PermissionConstants.PackedPermissionClaimType,
-                await rtoPCalcer.CalcPermissionsForUserAsync(userId)));
-            //and the same for the DataKey
-            claims.Add(new Claim(DataAuthConstants.HierarchicalKeyClaimName,
-                dataKeyCalc.CalcDataKeyForUser(userId)));
+                await rtoPCalcer.CalcPermissionsForUserAsync(context.Principal.Claims.GetUserIdFromClaims())));
 
             //Build a new ClaimsPrincipal and use it to replace the current ClaimsPrincipal
             var identity = new ClaimsIdentity(claims, "Cookie");
