+++ ServiceLayer/CodeCalledInStartup/AddClaimsToCookie.cs
@@ -5,7 +5,7 @@ using System;
 using CommonCache;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.DependencyInjection;
-using ServiceLayer.AuthCookieVersions;
+using ServiceLayer.AuthoriseSetup;
 using ServiceLayer.UserImpersonation.Concrete.Internal;
 
 namespace ServiceLayer.CodeCalledInStartup
