+++ ServiceLayer/CodeCalledInStartup/AddPermissionsToUserClaims.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using DataAuthorize;
+using DataKeyParts;
 using DataLayer.EfCode;
 using FeatureAuthorize;
 using Microsoft.AspNetCore.Identity;
